restructure and rename
