+++ src/app/components/CountDown/index.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import countdown from 'countdown';
-import c from '../../utils/c';
+import classnames, { withPrefix } from '#/utils/classnames';
 import './index.scss';
 
 export default
 class CountDown extends React.Component {
     static defaultProps = {
-        className: 'count-down'
+        prefix: 'count-down'
     }
 
     state = {}
@@ -22,12 +22,18 @@ class CountDown extends React.Component {
     render() {
         if (!this.state.timespan) return null;
         const { timespan } = this.state;
-        const arr = [['years', '年'], ['months', '月'], ['days', '日'], ['hours', '小时'], ['minutes', '分'], ['seconds', '秒']].map(([key, value]) => [key, value, timespan[key]]);
+        const arr = [['years', '年'], ['months', '月'], ['days', '日'], ['hours', '小时'], ['minutes', '分'], ['seconds', '秒']].map(([key, display]) => [key, display, timespan[key]]);
+
+        const p = withPrefix(this.props.prefix);
 
         return (
-            <div className={c(this.props.className)}>
+            <div className={classnames(p(), this.props.className)}>
                 {
-                    arr.map(([key, display, value], index) => <div key={key} data-show={arr.slice(0, index + 1).some(item => item[2])}>{value}<span>{display}</span></div>)
+                    arr.map(([key, display, value], index) => (
+                        <div className={p(key)} key={key} data-show={arr.slice(0, index + 1).some(item => item[2])}>
+                            {value}<span className={p('display-unit')}>{display}</span>
+                        </div>
+                    ))
                 }
             </div>
         );
