+++ src/app/containers/NavBar/index.scss
