+++ src/app/containers/NavItems/Live/Card/index.scss
