+++ src/app/containers/NavItems/Live/List/index.scss
