+++ src/app/containers/NavItems/News/Card/index.scss
