+++ src/app/containers/NavItems/News/List/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import c from '../../../../utils/c';
+import c from '../../../../utils/classnames';
 import './index.scss';
 
 export default
