+++ src/app/containers/NavItems/News/List/index.scss
