+++ src/app/containers/NavItems/News/index.scss
