+++ src/app/containers/NavLogo/index.scss
